added friends page to navigation bar and updated navbar to stay fixed on top of screen
