+++ study-connect/app/components/ClassesSidebar.tsx
@@ -75,7 +75,7 @@ export default function ClassesSidebar({ setSelectedClassId, setSelectedClassQua
   }, {} as Record<string, JoinedClass[]>);
 
   return (
-    <div className="w-64 bg-white shadow-md sticky top-0 h-1">
+    <div className="w-64 bg-white shadow-md h-1">
       <div className="p-4 border-b">
         <h2 className="text-lg font-semibold text-gray-800">My Classes</h2>
       </div>
